introduce failing test; remove LEMOP::Copy; move stress test for MatchTag from eval.rs to mod.rs
